+++ src/integration-test/resources/storage/complete.groovy
@@ -1,0 +1,5 @@
+databaseChangeLog = {
+  include file: 'storage/baseline.groovy'
+  include file: 'wtk/owned-file-lob.feat.groovy'
+  include file: 'wtk/stored-s3-object.feat.groovy'
+}