+++ src/integration-test/resources/storage/without-lob.groovy
@@ -1,0 +1,4 @@
+databaseChangeLog = {
+  include file: 'storage/baseline.groovy'
+  include file: 'wtk/stored-s3-object.feat.groovy'
+}