+++ src/integration-test/resources/storage/without-s3.groovy
@@ -1,0 +1,4 @@
+databaseChangeLog = {
+  include file: 'storage/baseline.groovy'
+  include file: 'wtk/owned-file-lob.feat.groovy'
+}